refactor: colocate misplaced tests, adopt .scenario. naming convention
Move 8 misplaced test files to sit next to their source modules.
Rename 6 scenario/integration tests with a .scenario. segment so
the filename reads as <module>.scenario.<topic>.test.ts — making it
clear the file tests a specific scenario of that module rather than
being the primary unit test.

Moves:
- labs.test.ts → core/store/ (tests store/labs.ts)
- validation.test.ts → bin-designer/utils/ (tests utils/validation.ts)
- useDesignerRouting.test.ts → hooks/ (tests hooks/useDesignerRouting.ts)
- replicadBin.integration.test.ts → worker/generators/
- stlExporter.test.ts → export/
- threemfExporter.test.ts → export/
- bridge.test.ts → generation/bridge/GenerationBridge.test.ts
- useBinInspector.test.ts → bin-inspector/hooks/
- designerValidation.test.ts → api/lib/

Scenario renames:
- half-bin-mode.test.ts → layout.scenario.half-bin.test.ts
- multi-bin.test.ts → layout.scenario.multi-bin.test.ts
- layout-displacement.test.ts → layout.scenario.displacement.test.ts
- inspirationGallery.test.ts → generateSuggestions.scenario.gallery.test.ts
- edgeCases.test.ts → replicadBin.scenario.edge-cases.test.ts
- compartmentActions.test.ts → designer.scenario.compartments.test.ts
- insertActions.test.ts → designer.scenario.inserts.test.ts

Also renamed meshCache.test.ts → meshCacheManager.test.ts to match
its actual source file.
