+++ src/core/store/layout.scenario.half-bin.test.ts
