+++ src/core/store/layout.scenario.multi-bin.test.ts
