+++ src/features/bin-designer/store/designer.scenario.compartments.test.ts
