+++ src/features/bin-designer/store/designer.scenario.inserts.test.ts
