+++ src/features/bin-designer/utils/validation.test.ts
