+++ src/features/bin-inspector/hooks/useBinInspector.test.ts
