+++ src/features/generation/worker/generators/replicadBin.integration.test.ts
